Initialize Git repo
The project was originally started as a small testing app that I did not plan on moving to Github.
Now that I have spent a bit more time developing it into a cohesive thing rather than a little code
playground, I have decided to start a Git repo. Sorry for the lack of commit history.